requirements.txt
streamlit
python-dotenv
pypdf
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
tokenizers==0.20.3
